Update opentextfile-method.md: fix code sample and another minor amendment (#91)
* Update opentextfile-method.md: correct code sample

Correct the code sample so the ForAppending constant has the correct value. The "iomode" table above is correct that the appending constant's value is 8; the code sample is wrong. (And make code formatting consistent while there.)

(This was originally fixed in commit b1ff695ad8ead6a92fea1c7def978cdccf709db3 in the old OfficeDev/VBA-content repository: https://github.com/OfficeDev/VBA-content/pull/548/commits. Somehow it was lost in subsequent merges for this file.)

* Update opentextfile-method.md: improve description

Note in the method's summary that write capabilities are present, to match the rest of the document's content.