Merge master to live 9-7 (#136)
* Fixed table formatting

* migrate old PRs from VBA-content to VBA-docs

* Clarification added to return values table.

* Updated links

* Lowercased enumeration

* Fixed links

* Changed "object", "property", "method" to lowercase

* Updated capitalization

* Updated date

* Added links

* Made headings level 2

* Changed heading level

* Changed "(./" to "("

The use of ./ for a link isn't needed and sometimes the links don't render correctly on the page. Removed this string from the link. All occurred in Access and Excel API files.

* Update replace-function.md (#86)

Extended the Description to make it clear, that the returned string can be a substring of the passed string expression. 
Modified the Part Description for the parameter start to reflect this.

Omitting these has caused a lot of confusion, as only the paragraph Remarks explains this.

* Update let-statement.md (#89)

Add missing spaces.

* Add missing spaces to set-statement.md (#90)

* Update opentextfile-method.md: fix code sample and another minor amendment (#91)

* Update opentextfile-method.md: correct code sample

Correct the code sample so the ForAppending constant has the correct value. The "iomode" table above is correct that the appending constant's value is 8; the code sample is wrong. (And make code formatting consistent while there.)

(This was originally fixed in commit b1ff695ad8ead6a92fea1c7def978cdccf709db3 in the old OfficeDev/VBA-content repository: https://github.com/OfficeDev/VBA-content/pull/548/commits. Somehow it was lost in subsequent merges for this file.)

* Update opentextfile-method.md: improve description

Note in the method's summary that write capabilities are present, to match the rest of the document's content.

* Update Excel.Application.OnUndo.md (#96)

Replace question marks with apostrophe.

* Correct Excel.Application.OnRepeat.md punctuation (#97)

* Update rnd-function.md (#112)

IntelliSense displays the parameter as Number.

Without going into a lengthy discussion, it should be indicated, that the values are picked from a list, thus are not random, only appears to be, thus are pseudo-random.

Corrected spacing and line breaks.

* Changed "About the contributor" headings to level 3 (#117)

Also deleted bookmarks like `<a name="AboutContributors"> </a>`.

I am happy to postpone this pull request if it interrupts other plans to fix formatting.

This pull request includes 28 Excel articles and 53 Access articles.

* Removed redundant "Enumeration"

* Update format-function-visual-basic-for-applications.md (#135)

* Update format-function-visual-basic-for-applications.md

Corrected many typos and missing new lines.
Removed the oooo format string which doesn't exist in VBA.
Adjusted or extended several explanations.
Adjusted casing of parameters to reflect the actual casing displayed by IntelliSense.

* Edit pass on updates

* Fixed backslash formatting

* Removed duplicate description

* Removed duplicate description

* Lowercased "members"