Update private-statement.md
The `Private` keyword is flat-out illegal inside a procedure scope. Also, the remark about having a wall-of-declarations at the start of a procedure, is actively harmful - variables should be declared close to where they're used, like in any other language; [Rubberduck](https://github.com/rubberduck-vba/Rubberduck) OSS VBIDE add-in has a refactoring feature to *move variable declaration closer to usage*.
`Dim` statements aren't executable, so their position in a procedure is irrelevant, so long as they appear before they are used.
Since this isn't a page about `Dim` or best practices in general, I held back from adding a note to that effect, but for correctness' sake I _had_ to remove the bit about locals being "generally declared near the top of a procedure".