Change filename and title
As the article describes the Dictionary objects Remove method and not that of the FileSystemObject the title should be updated and the filename changed. I am however uncertain if this is the right way to do this. This is the page that is shown when pressing the edit key in Microsoft Docs so if that relies on the  ms.assetid: dc895fae-17aa-4c51-4a35-8c3d3fd0e6fc or vblr6.chm2181952 I guess this would be Ok.