Code:'+' used when '&' should have been used
@Linda-Editor asked me to investigate the use of the '&amp;' text in the code example here. Have discovered that it is a mistake.

Whilst doing the investigation, I discovered another issue with the code. The operator '+' has been used when '&' should have been used. This file commit / PR is in respect to this latter issue. I decided to PR this issue as @Linda-Editor may make a coding mistake if she manually makes this change - it doesn't appear that she is that familiar with coding.