Misc. coding mistakes in code examples
@Linda-Editor asked me to investigate the use of the '&amp;' text in the code examples here. Have discovered that it is a mistake. See ISSUE #1029 for more info.

Whilst doing the investigation, I discovered other issues with the code. This file commit / PR is in respect to fixing these latter issues. I decided to PR this issue as @Linda-Editor may make a coding mistake if she manually makes the respective changes - it doesn't appear that she is that familiar with coding.