Clarify connection between Range.Item and cell access on Range(object) and add sections for Rows and Columns
So far, the documentation was misleading in the way that it suggested that Cells had two parameters and yielded single cells. The change in documentation clarifies that the single item access happens via Item and that the use of Cells is to convert the range into one consisting of single cells to achieve the intended behaviour of Item.

To contrast this, examples for Rows and Columns were added; for these, the behaviour of Item and thus also of the default member _Default is completely different.