Update instrrev-function.md
InstrRev does not return 0 if  start > Len(stringmatch) - it returns 0 if start > Len(stringcheck) - easy enough to prove: InstrRev("TEST", "E", 4) will return a value of 2, not zero, even though start > Len("E"). But InstrRev("TEST", "E", 5) does return 0, because now start > Len("TEST")

Much more subtle behavior is InstrRev will not find an instance of stringmatch unless the position of the end character of stringmatch is less than or equal to start. An example demonstrates: InstrRev("THIS IS A TEST STRING", "TEST", 13) returns a value of 0, but InstrRev("THIS IS A TEST STRING", "TEST", 14) returns a value of 11.