wdWrapFront=3. Change the topic or the enum?
wdWrapFront = 3, not 6, so I edited the topic. (I'm running Microsoft Office Professional 2016.) 

However, should wdWrapFront = 6 and should the enum code be changed? Is this a bug? 

If the enum code were changed, Word wouldn't be affected, because both 3 and 6 work the same. To test, I stepped through the following code. 

```vb
Sub TestShapeWrapFormatTypeValues()
    Dim shp As Shape
    Set shp = ActiveDocument.Shapes(1)
    shp.WrapFormat.Type = 1
    shp.WrapFormat.Type = 3
    shp.WrapFormat.Type = 1
    shp.WrapFormat.Type = 6
    shp.WrapFormat.Type = 1
End Sub
```