Update Excel.OLEObjects.md
The sentence: "When you first add a control to a sheet, the shape name and code name match. However, if you change either the shape name or code name, the other is not automatically changed to match."
seems to be no longer true with newer Excel versions. I was unable to get the Name and the Code Name inconsistent by any means.