Update PowerPoint.Presentation.DefaultLanguageID.md
The opening line is singularly misleading until clarified later. And then when clarified later it refers to "new presentations" but this is not what is happening with the current M365 beta channel release I have tested with (Version 2212 Build 16.0.15917.20000). Oddly, it does affects existing presentations after they are opened, which also contradicts this article.