To show users about Borders object when applied over a range of cells
The original issue shows the use of Borders object over a cell. It was left to users to find out on their own how Borders behave with the range of cells. This can be pretty confusing to new learners. I tried to see how it behaves with the range of cells. The output was not what i predicted. I thought Borders object will print borders only on the outside of the range of cells since its mentioned that "Use the Borders property to return the Borders collection, which contains all four borders" in the Remarks section. Those four borders are : xlEdgeBottom, xlEdgeLeft, xlEdgeRight, xlEdgeTop. But what i saw was it applied borders on the inside as well using xlInsideHorizontal and xlInsideVertical. So it returned 6 borders. 

So if we show how Borders applies to a range of cells, then new learners do not have to assume anything. It will make this article more clear and complete