Update work-with-shapes-drawing-objects.md
Corrected a typographical error in the article where the sentence fragment "and adds text to each shape that is an AutoShape." was repeated so that the text appeared (verbatim) like so:

"... and adds text to each shape that is an AutoShape. and adds text to each shape that is an AutoShape."

As shown above, rather than ending with a period, the paragraph in the article ended with a redundant string of text identical to the latter segment of the sentence immediately preceding it.  This was quite obviously included inadvertently.